[#179] [#180] Replaced (get|set) (votelist|vnlist|wishlist) with (get|set) ulist

Closes #180 and #179

See merge request thomas_herbeth/VNDBA!118

+++ app/src/main/java/com/booboot/vndbandroid/model/vndbandroid/VnlistData.kt
@@ -4,5 +4,5 @@ import com.booboot.vndbandroid.model.vndb.AccountItems
 
 data class VnlistData(
     val tabs: List<HomeTab>,
-    val items: Map<Int, AccountItems>
+    val items: Map<Long, AccountItems>
 )
